Add WorkManager CoroutineWorker
