+++ app/src/main/java/com/example/cryptoapp/domain/CoinRepository.kt
@@ -8,5 +8,5 @@ interface CoinRepository {
 
     fun getCoinInfo(fromSymbol: String): LiveData<CoinInfo>
 
-    suspend fun loadData()
+    fun loadData()
 }
